+++ google2/src/test/scala/org/broadinstitute/dsde/workbench/google2/mock/FakeGoogleComputeService.scala
@@ -2,9 +2,9 @@ package org.broadinstitute.dsde.workbench.google2.mock
 
 import cats.effect.IO
 import cats.mtl.Ask
-import com.google.cloud.compute.v1.{Firewall, Instance, MachineType, Network, Operation, Subnetwork, Tags, Zone}
+import com.google.api.gax.longrunning.OperationFuture
+import com.google.cloud.compute.v1._
 import org.broadinstitute.dsde.workbench.google2.{
-  ComputePollOperation,
   DeviceName,
   DiskName,
   FirewallRuleName,
@@ -22,13 +22,13 @@ import org.broadinstitute.dsde.workbench.model.google.GoogleProject
 class FakeGoogleComputeService extends GoogleComputeService[IO] {
   override def createInstance(project: GoogleProject, zone: ZoneName, instance: Instance)(implicit
     ev: Ask[IO, TraceId]
-  ): IO[Option[Operation]] =
-    IO.pure(Some(Operation.newBuilder().setId(123).setName("opName").setTargetId(258165385).build()))
+  ): IO[Option[OperationFuture[Operation, Operation]]] =
+    IO.pure(Some(new FakeComputeOperationFuture))
 
   override def deleteInstance(project: GoogleProject, zone: ZoneName, instanceName: InstanceName)(implicit
     ev: Ask[IO, TraceId]
-  ): IO[Option[Operation]] =
-    IO.pure(Some(Operation.newBuilder().setId(123).setName("opName").setTargetId(258165385).build()))
+  ): IO[Option[OperationFuture[Operation, Operation]]] =
+    IO.pure(Some(new FakeComputeOperationFuture))
 
   override def getInstance(project: GoogleProject, zone: ZoneName, instanceName: InstanceName)(implicit
     ev: Ask[IO, TraceId]
@@ -36,11 +36,11 @@ class FakeGoogleComputeService extends GoogleComputeService[IO] {
 
   override def stopInstance(project: GoogleProject, zone: ZoneName, instanceName: InstanceName)(implicit
     ev: Ask[IO, TraceId]
-  ): IO[Operation] = IO.pure(Operation.newBuilder().setId(123).setName("opName").setTargetId(258165385).build())
+  ): IO[OperationFuture[Operation, Operation]] = IO.pure(new FakeComputeOperationFuture)
 
   override def startInstance(project: GoogleProject, zone: ZoneName, instanceName: InstanceName)(implicit
     ev: Ask[IO, TraceId]
-  ): IO[Operation] = IO.pure(Operation.newBuilder().setId(123).setName("opName").setTargetId(258165385).build())
+  ): IO[OperationFuture[Operation, Operation]] = IO.pure(new FakeComputeOperationFuture)
 
   override def modifyInstanceMetadata(
     project: GoogleProject,
@@ -48,12 +48,12 @@ class FakeGoogleComputeService extends GoogleComputeService[IO] {
     instanceName: InstanceName,
     metadataToAdd: Map[String, String],
     metadataToRemove: Set[String]
-  )(implicit ev: Ask[IO, TraceId]): IO[Unit] =
-    IO.unit
+  )(implicit ev: Ask[IO, TraceId]): IO[Option[OperationFuture[Operation, Operation]]] =
+    IO.pure(Some(new FakeComputeOperationFuture))
 
   override def addFirewallRule(project: GoogleProject, firewall: Firewall)(implicit
     ev: Ask[IO, TraceId]
-  ): IO[Operation] = IO.pure(Operation.newBuilder().setId(123).setName("opName").setTargetId(258165385).build())
+  ): IO[OperationFuture[Operation, Operation]] = IO.pure(new FakeComputeOperationFuture)
 
   override def getFirewallRule(project: GoogleProject, firewallRuleName: FirewallRuleName)(implicit
     ev: Ask[IO, TraceId]
@@ -63,8 +63,8 @@ class FakeGoogleComputeService extends GoogleComputeService[IO] {
                               zone: ZoneName,
                               instanceName: InstanceName,
                               machineType: MachineTypeName
-  )(implicit ev: Ask[IO, TraceId]): IO[Unit] =
-    IO.unit
+  )(implicit ev: Ask[IO, TraceId]): IO[OperationFuture[Operation, Operation]] =
+    IO.pure(new FakeComputeOperationFuture)
 
   override def getMachineType(project: GoogleProject, zone: ZoneName, machineTypeName: MachineTypeName)(implicit
     ev: Ask[IO, TraceId]
@@ -76,7 +76,7 @@ class FakeGoogleComputeService extends GoogleComputeService[IO] {
 
   override def deleteFirewallRule(project: GoogleProject, firewallRuleName: FirewallRuleName)(implicit
     ev: Ask[IO, TraceId]
-  ): IO[Unit] = IO.unit
+  ): IO[Option[OperationFuture[Operation, Operation]]] = IO.pure(Some(new FakeComputeOperationFuture))
 
   override def getNetwork(project: GoogleProject, networkName: NetworkName)(implicit
     ev: Ask[IO, TraceId]
@@ -84,7 +84,7 @@ class FakeGoogleComputeService extends GoogleComputeService[IO] {
 
   override def createNetwork(project: GoogleProject, network: Network)(implicit
     ev: Ask[IO, TraceId]
-  ): IO[Operation] = IO.pure(Operation.newBuilder().setId(123).setName("opName").setTargetId(258165385).build())
+  ): IO[OperationFuture[Operation, Operation]] = IO.pure(new FakeComputeOperationFuture)
 
   override def getSubnetwork(project: GoogleProject, region: RegionName, subnetwork: SubnetworkName)(implicit
     ev: Ask[IO, TraceId]
@@ -92,35 +92,29 @@ class FakeGoogleComputeService extends GoogleComputeService[IO] {
 
   override def createSubnetwork(project: GoogleProject, region: RegionName, subnetwork: Subnetwork)(implicit
     ev: Ask[IO, TraceId]
-  ): IO[Operation] = IO.pure(Operation.newBuilder().setId(123).setName("opName").setTargetId(258165385).build())
+  ): IO[OperationFuture[Operation, Operation]] = IO.pure(new FakeComputeOperationFuture)
 
   def detachDisk(project: GoogleProject, zone: ZoneName, instanceName: InstanceName, deviceName: DeviceName)(implicit
     ev: Ask[IO, TraceId]
-  ): IO[Option[Operation]] =
-    IO.pure(Some(Operation.newBuilder().setId(123).setName("opName").setTargetId(258165385).build()))
+  ): IO[Option[OperationFuture[Operation, Operation]]] =
+    IO.pure(Some(new FakeComputeOperationFuture))
 
   override def deleteInstanceWithAutoDeleteDisk(
     project: GoogleProject,
     zone: ZoneName,
     instanceName: InstanceName,
     autoDeleteDisks: Set[DiskName]
-  )(implicit ev: Ask[IO, TraceId], computePollOperation: ComputePollOperation[IO]): IO[Option[Operation]] =
+  )(implicit ev: Ask[IO, TraceId]): IO[Option[OperationFuture[Operation, Operation]]] =
     IO.pure(
       Some(
-        Operation
-          .newBuilder()
-          .setId(123)
-          .setName("opName")
-          .setStatus(Operation.Status.DONE)
-          .setTargetId(258165385)
-          .build()
+        new FakeComputeOperationFuture
       )
     )
 
   override def setInstanceTags(project: GoogleProject, zone: ZoneName, instanceName: InstanceName, tags: Tags)(implicit
     ev: Ask[IO, TraceId]
-  ): IO[Operation] =
-    IO.pure(Operation.newBuilder().setId(123).setName("opName").setTargetId(258165385).build())
+  ): IO[OperationFuture[Operation, Operation]] =
+    IO.pure(new FakeComputeOperationFuture)
 }
 
 object FakeGoogleComputeService extends FakeGoogleComputeService